day 6

+++ adventofcode2023/day0/Solver.cs
@@ -8,12 +8,12 @@ namespace adventofcode2023.day0
 {
     public class Solver
     {
-        public int Part1(string[] lines)  
+        public long Part1(string[] lines)  
         {
             return 0;
         }
         
-        public int Part2(string[] lines)  
+        public long Part2(string[] lines)  
         {
             return 0;
         }
